# frontend/src/services/openapi.yml
openapi: 3.0.2
info:
  title: Microservices API
  description: API per l'autenticazione, gestione magazzino, ordini, pagamenti e spedizioni
  version: 1.0.0
  contact:
    name: Daniele Zoccali
    email: daniele.zoccali@gmail.com
servers:
  - url: https://microservices-alb-216051693.us-east-1.elb.amazonaws.com
    description: Server principale (ALB)
  - url: https://localhost:3002
    description: Server di sviluppo per Warehouse
  - url: https://localhost:3005
    description: Server di sviluppo per Payments
security:
  - bearerAuth: [] #"Per default, tutte le rotte richiedono un token JWT nel campo Authorization: Bearer <token>. "
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500.
        error:
          description: "Descrizione dell'errore; può essere una stringa o un oggetto."
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500
        request_id:
          type: string
          maxLength: 100
          description: "Identificativo della richiesta (se presente)."
          pattern: '^[0-9a-f]{32}$'
          example: "fd723645da950a54df847147b19f1108"
        success:
          type: boolean
          description: "Flag che, in alcuni casi, è restituito anche in risposta di errore."
...
    UploadError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500
    ErrorAPIData:
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
          example: "Errore nel recupero dell'utente"
          maxLength: 500
          pattern: ^.{1,500}$
        error:
          type: object
          additionalProperties: false
          properties:
            stringValue:
              type: string
              maxLength: 50
              pattern: ^.{1,50}$
              example: '":userId"'
            valueType:
              type: string
              maxLength: 50
              pattern: ^.{1,50}$
              example: "string"
            kind:
              type: string
              maxLength: 50
              pattern: ^.{1,50}$
              example: "ObjectId"
            value:
              type: string
              maxLength: 50
              pattern: ^.{1,50}$
              example: ":userId"
            path:
              type: string
              maxLength: 50
              pattern: ^.{1,50}$
              example: "_id"
            reason:
              type: object
              additionalProperties: false
              description: Dettagli opzionali sull'errore
            name:
              type: string
              maxLength: 50
              pattern: ^.{1,50}$
              example: "CastError"
            message:
              type: string
              maxLength: 500
              pattern: ^.{1,500}$
              example: "Cast to ObjectId failed for value \":userId\" (type string) at path \"_id\" for model \"User\""
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          pattern: '^\\S+@\\S+\\.\\S+$'
          example: poc@admin.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$'
          example: Abcdef1234!
    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - _id
        - username
        - email
        - role
        - isAdmin
        - token
      properties:
        _id:
          type: string
          maxLength: 24
          pattern: '^[0-9a-f]{24}$' # stringa esadecimale di 24 caratteri
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]{3,50}$'
        email:
          type: string
          format: email
          pattern: '^\S+@\S+\.\S+$'
          maxLength: 255
        role:
          type: string
          maxLength: 100
          pattern: '^.{1,100}$'
        isAdmin:
          type: boolean
        token:
          type: string
          maxLength: 2000
          pattern: '^[A-Za-z0-9\-_]+\.[A-Za-z0-9\-_]+\.[A-Za-z0-9\-_]+$'
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]{3,50}$'
        email:
          type: string
          format: email
          maxLength: 255
          pattern: '^\\S+@\\S+\\.\\S+$'
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$'
    RegisterResponse:
      type: object
      additionalProperties: false
      required:
        - _id
        - username
        - email
        - role
        - token
      properties:
        _id:
          type: string
          maxLength: 24
          pattern: '^[0-9a-f]{24}$' # stringa esadecimale di 24 caratteri
        username:
          type: string
          pattern: '^[a-zA-Z0-9_-]{3,50}$'
          maxLength: 50
        email:
          type: string
          format: email
          pattern: '^\S+@\S+\.\S+$'
          maxLength: 255
        role:
          type: string
          enum: [admin, user] # oppure un pattern, se i ruoli possono variare
        token:
          type: string
          maxLength: 2000
          pattern: '^[A-Za-z0-9\-_]+\.[A-Za-z0-9\-_]+\.[A-Za-z0-9\-_]+$'
    UploadProfileResponse:
      type: object
      additionalProperties: false
      required:
        - message
        - user
      properties:
        message:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500.
        user:
          type: object
          additionalProperties: false
          required: # anche profileImage in teoria ma teniamoci flessibili
            - username
            - email
            - _id
          properties:
            username:
              type: string
              maxLength: 50
              pattern: '^[a-zA-Z0-9_-]{1,50}$'
            email:
              type: string
              format: email
              maxLength: 255
              pattern: '^\S+@\S+\.\S+$'
            profileImage: #(opzionale)
              type: string
              format: uri
              maxLength: 500
              pattern: '^https:\/\/.*$'
            _id:
              type: string
              maxLength: 24
              pattern: '^[0-9a-f]{24}$'
    ApiDataResponse:
      type: object
      additionalProperties: false
      required:
        - _id
        - username
        - email
      properties:
        _id:
          type: string
          maxLength: 24
          pattern: '^[0-9a-f]{24}$'
        username:
          type: string
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]{1,50}$'
        email:
          type: string
          format: email
          maxLength: 255
          pattern: '^\S+@\S+\.\S+$'
...
    ProfileResponse:
      type: object
      additionalProperties: false
      required:
        - user
      properties:
        user:
          type: object
          additionalProperties: false
          required:
            - _id
            - username
            - email
            - role
            - createdAt
            - updatedAt
            - __v
          properties:
            _id:
              type: string
              maxLength: 24
              pattern: '^[0-9a-f]{24}$'
            username:
              type: string
              maxLength: 50
              pattern: '^[a-zA-Z0-9_-]{1,50}$'
            email:
              type: string
              format: email
              maxLength: 255
              pattern: '^\S+@\S+\.\S+$'
            role:
              type: string
              maxLength: 20
              pattern: '^[a-z]+$'
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            __v:
              type: integer
            profileImage: #(opzionale)
              type: string
              format: uri
              maxLength: 500
              pattern: '^https:\/\/.*$'
    ProfileUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          maxLength: 24
          pattern: '^[0-9a-f]{24}$' #stringa esadecimale di 24 caratteri
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]{3,50}$'
        email:
          type: string
          format: email
          maxLength: 255
          pattern: '^\\S+@\\S+\\.\\S+$'
        imageKey:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$'
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$'
    
    #Ok (200)
    UploadResponse:
      type: object
      additionalProperties: false
      required:
        - imageKey
      properties:
        imageKey:
          type: string
          maxLength: 100
          pattern: '^.{1,100}$'
...
    ErrorOnlyMessage:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500.
...
    ErrorRegister:
      type: object
      additionalProperties: false
      required:
        - message
        - error
      properties:
        message:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500.
        error:  
          type: string
          maxLength: 1000
          pattern: '^.{1,1000}$'
    
     # anche qui limitiamo la lunghezza ma accettiamo qualsiasi contenuto testuale

 # accetta qualsiasi stringa con almeno 1 carattere fino a 100
paths:
...
  /api/auth/login:
    servers:
    - url: https://microservices-alb-216051693.us-east-1.elb.amazonaws.com
      description: Server principale
    post:
      summary: Accesso utente
      tags:
        - Authentication
      security: [] #Override, non serve sicurezza (mezzo per ottenere il token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login riuscito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Credenziali non valide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
              examples: #Serve a non fare fallire il test in 42crunch 
                invalidCredentials:
                  summary: Invalid credentials
                  value:
                    message: "Credenziali non valide" #Kong gateway message

        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
#Ok Happy path in teoria
  /api/auth/register:
    servers:
    - url: https://microservices-alb-216051693.us-east-1.elb.amazonaws.com
      description: Server principale
    post:
      summary: Registrazione nuovo utente
      tags:
        - Authentication
      security: [] #Override, non serve sicurezza (mezzo per ottenere il token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              validRegistration:
                summary: ValidRegistration
                value:
                  username: TestUserPoC
                  email: test@poc42.com
                  password: Abcdef12345!     
      responses:
        '201':
          description: Utente registrato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Dati di registrazione non validi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email già esistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRegister'

        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/profile:
    servers:
    - url: https://microservices-alb-216051693.us-east-1.elb.amazonaws.com
      description: Server principale
    get:
      summary: Ottieni profilo utente
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profilo utente recuperato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Non hai i permessi necessari
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Profilo non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/html:
              schema:
                type: string
                maxLength: 4096
                pattern: '^<html>[\\s\\S]{1,4090}</html>$'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/upload:
    servers:
    - url: https://microservices-alb-216051693.us-east-1.elb.amazonaws.com
      description: Server principale
    post:
      summary: Carica file
      tags:
        - Authentication
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              properties:
                file:
                  type: string
                  format: binary
                  maxLength: 16384 
      responses:
        '200':
          description: Upload completato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Errore nell'upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadError' #perchè restituisce "error" al posto di "message"
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadError'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadError'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadError'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadError'

#SKIP HAPPY PATH PROBLEMA S3 (TOKEN)
  /api/auth/uploadProfile/{id}:
    servers:
    - url: https://microservices-alb-216051693.us-east-1.elb.amazonaws.com
      description: Server principale
    put:
      summary: Aggiorna profilo utente
      tags:
        - Authentication
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$' #stringa esadecimale di 24 caratteri
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: Profilo aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadProfileResponse'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
              examples: #Serve a non fare fallire il test in 42crunch 
                missingAuthorization:
                  summary: Unauthorized
                  value:
                    message: "Unauthorized" #Kong gateway message
        '400':
          description: Errore nell'aggiornamento del profilo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples: #Serve a non fare fallire il test in 42crunch 
                badRequest:
                  summary: Bad Request
                  value:
                    message: "Bad Request: userId missing in URL" #Kong gateway message
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
              examples: #Serve a non fare fallire il test in 42crunch 
                missingAuthorization:
                  summary: Unauthorized
                  value:
                    message: "Unauthorized: Missing Authorization header" #Kong gateway message
                missingToken:
                  summary: MissingToken
                  value:
                    message: "Unauthorized: Invalid or missing token" #Kong gateway message
                invalidJWT:
                  summary: InvalidToken
                  value:
                    message: "Unauthorized: Invalid JWT" #Kong gateway message
                userId:
                  summary: InvalidUserId insided Token
                  value:
                    message: "Unauthorized: Missing user ID in token"
        '404':
          description: Utente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/html:
              schema:
                type: string
                maxLength: 4096
                pattern: '^<html>[\\s\\S]{1,4090}</html>$'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/data/{userId}:
    servers:
    - url: https://microservices-alb-216051693.us-east-1.elb.amazonaws.com
      description: Server principale
    get:
      summary: Ottieni dati utente
      tags:
        - Authentication
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$' #stringa esadecimale di 24 caratteri
      responses:
        '200':
          description: Dati utente recuperati
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDataResponse'
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Utente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/html:
              schema:
                type: string
                maxLength: 4096
                pattern: '^<html>[\\s\\S]{1,4090}</html>$'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAPIData'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...